+++ packages/navi-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @awell-health/navi-core
 
+## 0.0.8
+
+### Patch Changes
+
+- updates to navi-core type; add optional values to session schema
+
 ## 0.0.7
 
 ### Patch Changes