+++ packages/navi-js/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @awell-health/navi-js
 
+## 0.0.8
+
+### Patch Changes
+
+- updates to navi-core type; add optional values to session schema
+- Updated dependencies
+  - @awell-health/navi-core@0.0.8
+
 ## 0.0.7
 
 ### Patch Changes